Add alias targets for every generated target.
The alias target omits the version since we are doing one version per
artifact per repository. Instead of referring to
@maven//:android_arch_lifecycle_common_1_1_1 or
maven_artifact("android.arch.lifecycle:common:1.1.1"), specify
@maven//:android_arch_lifecycle_common directly.

This is a safe behavior because the versions are pinned in WORKSPACE's
maven_install. The aliases only switch targets when the dep versions in
maven_install are changed.

e.g.

alias(
  name = "android_arch_lifecycle_common",
  actual = "@maven//:android_arch_lifecycle_common_1_1_1",
)

java_import(
  name = "android_arch_lifecycle_common_1_1_1",
  jars = ["@maven//:https/maven.google.com/android/arch/lifecycle/common/1.1.1/common-1.1.1.jar"],
  srcjar = "@maven//:https/maven.google.com/android/arch/lifecycle/common/1.1.1/common-1.1.1-sources.jar",
  deps = ["@maven//:com_android_support_support_annotations_28_0_0"],
)